Fix icons bleeding and some atlas calculation inconsistencies

+++ rts/Rendering/Textures/MultiPageAtlasAlloc.hpp
@@ -130,7 +130,7 @@ public:
 
 		return (reqLevels == std::numeric_limits<int>::max()) ? 0 : reqLevels;
 	}
-	uint32_t GetNumPages() const override { return allocators.size(); }
+	uint32_t GetNumPages() const override { return static_cast<uint32_t>(allocators.size()); }
 private:
 	uint32_t maxPages;
 	std::vector<std::unique_ptr<AtlasAlloc>> allocators;
